apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "digital-barn-finds-api"
version = "0.1.0"
description = "API and ingestion service for Digital Barn Finds."
requires-python = ">=3.12"
dependencies = [
  "alembic==1.14.0",
  "beautifulsoup4==4.12.3",
  "fastapi==0.115.6",
  "gunicorn==23.0.0",
  "httpx==0.28.1",
  "psycopg[binary]==3.2.3",
  "pydantic-settings==2.7.0",
  "sqlalchemy==2.0.36",
  "uvicorn[standard]==0.32.1"
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.4",
  "ruff==0.8.4"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py312"

